move PublicClassMapper to our package (#2008)
